Add custom type JSONTime
Add custom type JSONTime not to depend on local time.

Because,
the default locale of `time.Time` is not "UTC" on unit testing in github actions workflow.
But if attribute of value in struct has `time.Time`, the locale of the value  is set to "UTC" after `json.Unmarshal`.
And then the unit tests depends on it are failed.